# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Vice Intelligence Platform — GTA Gaming CMS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/categories:
    get:
      operationId: listAdminCategories
      tags: [cms-categories]
      summary: List all categories
      responses:
        "200":
          description: Categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createAdminCategory
      tags: [cms-categories]
      summary: Create category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /admin/categories/{id}:
    patch:
      operationId: updateAdminCategory
      tags: [cms-categories]
      summary: Update category
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteAdminCategory
      tags: [cms-categories]
      summary: Delete category
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204":
          description: Deleted

  # ─── CMS TAGS ────────────────────────────────────────────────────────────────
  /admin/tags:
    get:
      operationId: listAdminTags
      tags: [cms-tags]
      summary: List all tags
      responses:
        "200":
          description: Tags
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tag"
    post:
      operationId: createAdminTag
      tags: [cms-tags]
      summary: Create tag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TagInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tag"
  /admin/tags/{id}:
    patch:
      operationId: updateAdminTag
      tags: [cms-tags]
      summary: Update tag
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TagUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tag"
    delete:
      operationId: deleteAdminTag
      tags: [cms-tags]
      summary: Delete tag
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204":
          description: Deleted

  # ─── CMS MEDIA ───────────────────────────────────────────────────────────────
  /admin/media:
    get:
      operationId: listAdminMedia
      tags: [cms-media]
      summary: List media library
      parameters:
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: limit, in: query, schema: { type: integer, default: 40 } }
        - { name: search, in: query, schema: { type: string } }
      responses:
        "200":
          description: Media items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaListResponse"
  /admin/media/{id}:
    patch:
      operationId: updateAdminMedia
      tags: [cms-media]
      summary: Update media metadata
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MediaUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaItem"
    delete:
      operationId: deleteAdminMedia
      tags: [cms-media]
      summary: Delete media item
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204":
          description: Deleted

  # ─── CMS SETTINGS ────────────────────────────────────────────────────────────
  /admin/settings:
    get:
      operationId: getAdminSettings
      tags: [cms-settings]
      summary: Get site settings
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteSettings"
    patch:
      operationId: updateAdminSettings
      tags: [cms-settings]
      summary: Update site settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SiteSettingsUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteSettings"

  # ─── CMS HOMEPAGE ────────────────────────────────────────────────────────────
  /admin/homepage-config:
    get:
      operationId: getAdminHomepageConfig
      tags: [cms-homepage]
      summary: Get homepage configuration
      responses:
        "200":
          description: Homepage config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HomepageConfig"
    patch:
      operationId: updateAdminHomepageConfig
      tags: [cms-homepage]
      summary: Update homepage configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HomepageConfigUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HomepageConfig"
components:
  schemas:
...
    Category:
      type: object
      required: [id, name, slug]
      properties:
        id:           { type: integer }
        name:         { type: string }
        slug:         { type: string }
        description:  { type: ["string", "null"] }
        color:        { type: ["string", "null"] }
        parentId:     { type: ["integer", "null"] }
        articleCount: { type: integer }
        createdAt:    { type: string }
    CategoryInput:
      type: object
      required: [name]
      properties:
        name:        { type: string, minLength: 1 }
        slug:        { type: string }
        description: { type: string }
        color:       { type: string }
        parentId:    { type: integer }
    CategoryUpdate:
      type: object
      properties:
        name:        { type: string, minLength: 1 }
        slug:        { type: string }
        description: { type: string }
        color:       { type: string }
        parentId:    { type: integer }

    # ── TAG SCHEMAS ──────────────────────────────────────────────────────────
    Tag:
      type: object
      required: [id, name, slug]
      properties:
        id:           { type: integer }
        name:         { type: string }
        slug:         { type: string }
        articleCount: { type: integer }
        createdAt:    { type: string }
    TagInput:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1 }
        slug: { type: string }
    TagUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        slug: { type: string }

    # ── MEDIA SCHEMAS ────────────────────────────────────────────────────────
    MediaItem:
      type: object
      required: [id, url, filename, createdAt]
      properties:
        id:        { type: integer }
        url:       { type: string }
        filename:  { type: string }
        mimeType:  { type: ["string", "null"] }
        size:      { type: ["integer", "null"] }
        width:     { type: ["integer", "null"] }
        height:    { type: ["integer", "null"] }
        alt:       { type: ["string", "null"] }
        caption:   { type: ["string", "null"] }
        createdAt: { type: string }
    MediaUpdate:
      type: object
      properties:
        alt:     { type: string }
        caption: { type: string }
    MediaListResponse:
      type: object
      required: [items, total, page, limit]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/MediaItem"
        total: { type: integer }
        page:  { type: integer }
        limit: { type: integer }

    # ── SETTINGS SCHEMAS ─────────────────────────────────────────────────────
    SiteSettings:
      type: object
      required: [siteName]
      properties:
        siteName:       { type: string }
        siteDescription: { type: ["string", "null"] }
        siteUrl:        { type: ["string", "null"] }
        logoUrl:        { type: ["string", "null"] }
        faviconUrl:     { type: ["string", "null"] }
        twitterHandle:  { type: ["string", "null"] }
        facebookUrl:    { type: ["string", "null"] }
        youtubeUrl:     { type: ["string", "null"] }
        discordUrl:     { type: ["string", "null"] }
        contactEmail:   { type: ["string", "null"] }
        googleAnalyticsId: { type: ["string", "null"] }
        adsenseId:      { type: ["string", "null"] }
        updatedAt:      { type: ["string", "null"] }
    SiteSettingsUpdate:
      type: object
      properties:
        siteName:        { type: string }
        siteDescription: { type: string }
        siteUrl:         { type: string }
        logoUrl:         { type: string }
        faviconUrl:      { type: string }
        twitterHandle:   { type: string }
        facebookUrl:     { type: string }
        youtubeUrl:      { type: string }
        discordUrl:      { type: string }
        contactEmail:    { type: string }
        googleAnalyticsId: { type: string }
        adsenseId:       { type: string }

    # ── HOMEPAGE SCHEMAS ─────────────────────────────────────────────────────
    HomepageConfig:
      type: object
      properties:
        heroTitle:       { type: ["string", "null"] }
        heroSubtitle:    { type: ["string", "null"] }
        heroImage:       { type: ["string", "null"] }
        featuredIds:     { type: array, items: { type: integer } }
        editorPickIds:   { type: array, items: { type: integer } }
        showTrending:    { type: boolean }
        showLatest:      { type: boolean }
        showPopular:     { type: boolean }
        showByGame:      { type: boolean }
        updatedAt:       { type: ["string", "null"] }
    HomepageConfigUpdate:
      type: object
      properties:
        heroTitle:       { type: string }
        heroSubtitle:    { type: string }
        heroImage:       { type: string }
        featuredIds:     { type: array, items: { type: integer } }
        editorPickIds:   { type: array, items: { type: integer } }
        showTrending:    { type: boolean }
        showLatest:      { type: boolean }
        showPopular:     { type: boolean }
        showByGame:      { type: boolean }
...