# lib/api-zod/src/generated/types/homepageConfigUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Vice Intelligence Platform — GTA Gaming CMS API
 * OpenAPI spec version: 0.2.0
 */

export interface HomepageConfigUpdate {
  heroTitle?: string;
  heroSubtitle?: string;
  heroImage?: string;
  featuredIds?: number[];
  editorPickIds?: number[];
  showTrending?: boolean;
  showLatest?: boolean;
  showPopular?: boolean;
  showByGame?: boolean;
}