# lib/api-zod/src/generated/types/mediaListResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Vice Intelligence Platform — GTA Gaming CMS API
 * OpenAPI spec version: 0.2.0
 */
import type { MediaItem } from './mediaItem';

export interface MediaListResponse {
  items: MediaItem[];
  total: number;
  page: number;
  limit: number;
}