# lib/api-zod/src/generated/types/siteSettings.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Vice Intelligence Platform — GTA Gaming CMS API
 * OpenAPI spec version: 0.2.0
 */

export interface SiteSettings {
  siteName: string;
  /** @nullable */
  siteDescription?: string | null;
  /** @nullable */
  siteUrl?: string | null;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  faviconUrl?: string | null;
  /** @nullable */
  twitterHandle?: string | null;
  /** @nullable */
  facebookUrl?: string | null;
  /** @nullable */
  youtubeUrl?: string | null;
  /** @nullable */
  discordUrl?: string | null;
  /** @nullable */
  contactEmail?: string | null;
  /** @nullable */
  googleAnalyticsId?: string | null;
  /** @nullable */
  adsenseId?: string | null;
  /** @nullable */
  updatedAt?: string | null;
}